divided adminView to attribute, attribute_group, building_class and session
